+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ﻿## Version 4.14 [In Development]
 
 **Retail:**  
-- Updated PermoksAccountManager profile (Font size 10 > 11)  
+- Added option to select a different ElvUI focus nameplate texture  
+- Updated PermoksAccountManager profile  
 - Updated SkironCooldownManager profile  
 
 ## Version 4.13 [13-05-2026]